+++ djangocms_installer/main.py
@@ -25,13 +25,14 @@ def execute():
                              'Please wait while I install dependencies\n')
             if not config_data.no_deps:
                 if config_data.requirements_file:
-                    install.requirements(config_data.requirements_file,
-                                         config_data.pip_options, True)
+                    install.requirements(
+                        config_data.requirements_file, config_data.pip_options, True
+                    )
                 else:
-                    install.requirements(config_data.requirements,
-                                         config_data.pip_options)
-            sys.stdout.write('Dependencies installed\n'
-                             'Creating the project')
+                    install.requirements(
+                        config_data.requirements, config_data.pip_options
+                    )
+            sys.stdout.write('Dependencies installed\nCreating the project')
             install.check_install(config_data)
             django.create_project(config_data)
             django.patch_settings(config_data)
@@ -40,7 +41,9 @@ def execute():
                 django.setup_database(config_data)
             if config_data.starting_page:
                 django.load_starting_page(config_data)
-            if config_data.aldryn:
+            if not config_data.requirements_file:
+                install.write_requirements(config_data)
+            if config_data.aldryn:  # pragma: no cover
                 sys.stdout.write('Project created!\n')
                 sys.stdout.write('aldryn boilerplate requires action before '
                                  'you can actually run the project.\n'